Add files via upload
A modified version adding DST for the UK, the source code has a USA version commented out, so needs the UK tests replaced by the US tests in the 'display_time' function.. It was not easy to work out the logic to get DST correct. Hours have not been implemented, but the error in change-over is no more than 1-hour.